requirements.txt
pandas
numpy
plotly
jinja2
MetaTrader5
pytz
python-dotenv